\name{restricted.betweenness-package}
\alias{restricted.betweenness-package}
\alias{restricted.betweenness}
\docType{package}
\title{
Restricted betweenness
}
\description{
Restricted betweenness calculates the betweenness centrality of every nodes in a graph using only shortests paths between a set of source nodes and a set of sink nodes 
License: What license is it under?
}
\details{
\tabular{ll}{
Package: \tab restricted.betweenness\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-07\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
